requirements.txt
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.34.0
azure-functions==1.23.0
azure-mgmt-core==1.5.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
msrest==0.7.1
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
